don't work; save MapExecutionTaskRendering subject orders format

+++ code/CCoA_ij_idea/src/Logic/FootprintSpaceTime/FootprintSpaceTime.java
@@ -5,14 +5,14 @@ import Logic.TypesInLevel;
 import java.util.List;
 
 public interface FootprintSpaceTime {
-    public List<PolygonExtended> getAreaFromWhen(PolygonExtended areaFind, long when);
-    public List<PolygonExtended> getAreaFromWhen(PolygonExtended areaFind, long when, TypesInLevel type);
+    public List<PolygonExtended> getAreaFromWhen(PolygonExtended areaFind, int when);
+    public List<PolygonExtended> getAreaFromWhen(PolygonExtended areaFind, int when, TypesInLevel type);
 
-    public void addPointRadius(int ID, PolygonExtended Place, long time);
+    public void addPointRadius(int ID, PolygonExtended Place, int time);
     public void addPointRadius(int ID, PolygonExtended Place);
     public void deleteFootprints(int ID);
 
-    public boolean getAccessPlace(PolygonExtended place, long time, TypesInLevel type);
+    public boolean getAccessPlace(PolygonExtended place, int time, TypesInLevel type);
 
     /*TODO: влажные мечты: здесь можно сделать получение утромбованной точки размещения полигона и узнавание важности точки*/
