+++ code/CCoA_ij_idea/src/Logic/FootprintSpaceTime/Point.java
@@ -1,8 +1,8 @@
 package Logic.FootprintSpaceTime;
 
 public interface Point {
-    public long getX();
-    public long getY();
-    public void setX(long x);
-    public void setY(long y);
+    public int getX();
+    public int getY();
+    public void setX(int x);
+    public void setY(int y);
 }
