+++ code/CCoA_ij_idea/src/Logic/Landscape/Landscape.java
@@ -9,7 +9,7 @@ import java.util.List;
 
 public interface Landscape {
     public void fillArea();
-    public long speedInfluenceEnvironmentOnProperties(long speed);
+    public int speedInfluenceEnvironmentOnProperties(int speed);
     public List<Point> getPointsDirectAccess(PolygonExtended area);
     public Path getDijkstraPath(Point from, Point to); //TODO: толстая машина не может проезать через узкое ущелье
 }
